# perf(beghilosz): ~25% increase