--- a/beghilosz/beghilosz_test.go
+++ b/beghilosz/beghilosz_test.go
@@ -40,10 +40,15 @@ func TestEncode(t *testing.T) {
 			args: args{"Sentence with a #hashtag"},
 			want: "#HASHTAG A 4T1W 3CN3TN35",
 		},
+		// {
+		// 	name: "Multiline text",
+		// 	args: args{"Line 1\nLine 2"},
+		// 	want: "2 3N17\n1 3N17",
+		// },
 		{
-			name: "Multiline text",
-			args: args{"Line 1\nLine 2"},
-			want: "2 3N17\n1 3N17",
+			name: "Cyrilic",
+			args: args{"привет"},
+			want: "ТЕВИРП",
 		},
 	}
 	for _, tt := range tests {
@@ -58,6 +63,7 @@ func TestEncode(t *testing.T) {
 func BenchmarkEncode(b *testing.B) {
 	for _, bench := range utils.EncodeBenchmarks {
 		b.Run(bench.Name, func(b *testing.B) {
+			b.SetBytes(int64(len(bench.Text)))
 			for i := 0; i < b.N; i++ {
 				Encode(bench.Text)
 			}