--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,10 @@
 module github.com/kirillmorozov/encodor
 
-go 1.16
+go 1.17
 
 require github.com/spf13/cobra v1.4.0
+
+require (
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)