### Benchmarks

Here are benchstat results when compared to 08ba9e7c26708120ed51e5b0f207d0b7d70212e9.

```
name                   old time/op    new time/op    delta
Encode/Multiline-10      1.13µs ± 1%    0.62µs ± 0%  -45.68%  (p=0.000 n=9+10)
Encode/Lorem_Ipsum-10    8.23µs ± 1%    8.01µs ± 1%   -2.77%  (p=0.000 n=10+10)
[Geo mean]               3.05µs         2.22µs       -27.33%

name                   old alloc/op   new alloc/op   delta
Encode/Multiline-10      1.15kB ± 0%    0.37kB ± 0%  -68.06%  (p=0.000 n=10+10)
Encode/Lorem_Ipsum-10    4.58kB ± 0%    3.92kB ± 0%  -14.34%  (p=0.000 n=10+10)
[Geo mean]               2.30kB         1.20kB       -47.69%

name                   old allocs/op  new allocs/op  delta
Encode/Multiline-10        22.0 ± 0%      12.0 ± 0%  -45.45%  (p=0.000 n=10+10)
Encode/Lorem_Ipsum-10       206 ± 0%       202 ± 0%   -1.94%  (p=0.000 n=10+10)
[Geo mean]                 67.3           49.2       -26.87%
```